Moved testGenericQueryCacheRoundTrip to generic_query_cache_test.go
